Update tradeBot
I cannot for the life of me figure out effective error handling in solidity, so for now the killCasino() function doesn't loop which is incredibly annoying.  I tried some try catch stuff, but that doesn't work, and I tried keeping track of how many times killCasino() runs before it reverts, but since the whole state reverts that counter gets reset.  I'm trying now to be calling killCasino() in a loop from a different contract, which hopefully would mean I can not have the whole transaction revert when it fails but just the last iteration, but that doesn't sound like the right way to do it.